+++ app/src/main/java/com/techcourse/controller/HomeController.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.techcourse.controller;
 
 import context.org.springframework.stereotype.Controller;
 import jakarta.servlet.http.HttpServletRequest;
