+++ app/src/test/java/com/techcourse/controller/LoginControllerTest.java
@@ -1,7 +1,5 @@
 package com.techcourse.controller;
 
-import com.techcourse.LoginController;
-import com.techcourse.UserSession;
 import com.techcourse.domain.User;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -10,8 +8,6 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import webmvc.org.springframework.web.servlet.ModelAndView;
 import webmvc.org.springframework.web.servlet.view.JspView;
-
-import java.util.Optional;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
